refactor: move bond reductions keeper methods to seperate file
